requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
pytest
matplotlib==3.9.2
openai==1.30.1
httpx==0.27.2